refactor: 예약 관련 서비스 테스트 추상 클래스 추가 및 클래스명 변경
