+++ src/test/java/roomescape/reservation/application/BookingQueryServiceTest.java
@@ -1,12 +1,9 @@
 package roomescape.reservation.application;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import roomescape.common.ServiceTest;
-import roomescape.member.application.MemberService;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
@@ -22,41 +19,16 @@ import static roomescape.TestFixture.MIA_RESERVATION_TIME;
 import static roomescape.TestFixture.TOMMY_NAME;
 import static roomescape.TestFixture.TOMMY_RESERVATION;
 import static roomescape.TestFixture.TOMMY_RESERVATION_TIME;
-import static roomescape.TestFixture.USER_MIA;
-import static roomescape.TestFixture.USER_TOMMY;
-import static roomescape.TestFixture.WOOTECO_THEME;
 import static roomescape.TestFixture.WOOTECO_THEME_NAME;
 import static roomescape.reservation.domain.ReservationStatus.BOOKING;
 
-class ReservationQueryServiceTest extends ServiceTest {
+class BookingQueryServiceTest extends ReservationServiceTest {
     @Autowired
     private BookingQueryService bookingQueryService;
 
     @Autowired
     @Qualifier("bookingManageService")
     private ReservationManageService bookingManageService;
-
-    @Autowired
-    private ReservationTimeService reservationTimeService;
-
-    @Autowired
-    private ThemeService themeService;
-
-    @Autowired
-    private MemberService memberService;
-
-    private ReservationTime miaReservationTime;
-    private Theme wootecoTheme;
-    private Member mia;
-    private Member tommy;
-
-    @BeforeEach
-    void setUp() {
-        this.miaReservationTime = reservationTimeService.create(new ReservationTime(MIA_RESERVATION_TIME));
-        this.wootecoTheme = themeService.create(WOOTECO_THEME());
-        this.mia = memberService.create(USER_MIA());
-        this.tommy = memberService.create(USER_TOMMY());
-    }
 
     @Test
     @DisplayName("모든 예약 목록을 조회한다.")
