+++ src/test/java/roomescape/reservation/application/ReservationServiceTest.java
@@ -1,0 +1,41 @@
+package roomescape.reservation.application;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import roomescape.common.ServiceTest;
+import roomescape.member.application.MemberService;
+import roomescape.member.domain.Member;
+import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.domain.Theme;
+
+import static roomescape.TestFixture.MIA_RESERVATION_TIME;
+import static roomescape.TestFixture.USER_ADMIN;
+import static roomescape.TestFixture.USER_MIA;
+import static roomescape.TestFixture.USER_TOMMY;
+import static roomescape.TestFixture.WOOTECO_THEME;
+
+abstract class ReservationServiceTest extends ServiceTest {
+    @Autowired
+    private ReservationTimeService reservationTimeService;
+
+    @Autowired
+    private ThemeService themeService;
+
+    @Autowired
+    private MemberService memberService;
+
+    protected ReservationTime miaReservationTime;
+    protected Theme wootecoTheme;
+    protected Member mia;
+    protected Member tommy;
+    protected Member admin;
+
+    @BeforeEach
+    void setUp() {
+        this.miaReservationTime = reservationTimeService.create(new ReservationTime(MIA_RESERVATION_TIME));
+        this.wootecoTheme = themeService.create(WOOTECO_THEME());
+        this.mia = memberService.create(USER_MIA());
+        this.tommy = memberService.create(USER_TOMMY());
+        this.admin = memberService.create(USER_ADMIN());
+    }
+}
